Move all time functions to mono-time.c
